src/ai/requirements.txt
tensorflow
keras
pillow
numpy
scikit-learn
